==> 0.Preparar_archivos_csv.ipynb <==
IMPORTAMOS LIBRERIAS

import pandas as pd
import json
import ast
import gzip

IMPORTAMOS LOS ARCHIVOS  JSON

#Abre el archivo y procesa cada línea
with open('Datasets\output_steam_games.json', 'r', encoding='utf-8') as archivo:
    data = [json.loads(line) for line in archivo]

# Convertir la lista de objetos JSON en un DataFrame
steam_gamesdf = pd.DataFrame(data)

#almacenamos en una lista
df_lista = []

# Ruta del archivo JSON
file_path = 'Datasets/australian_user_reviews.json'

# Abre el archivo y procesar cada línea
with open(file_path, 'r', encoding='utf-8') as file:
    for line in file:
        try:
            # Usar ast.literal_eval para convertir la línea en un diccionario
            json_data = ast.literal_eval(line)
            df_lista.append(json_data)
        except ValueError as e:
            print(f"Error en la línea: {line}")
            continue

user_reviewsdf = pd.DataFrame(df_lista)

#almacenamos en una lista
df_lista1 = []

# Ruta del archivo JSON
file_path = 'Datasets/australian_users_items.json'

# Abre el archivo y procesar cada línea
with open(file_path, 'r', encoding='utf-8') as file:
    for line in file:
        try:
            # Usar ast.literal_eval para convertir la línea en un diccionario
            json_data = ast.literal_eval(line)
            df_lista1.append(json_data)
        except ValueError as e:
            print(f"Error en la línea: {line}")
            continue

user_itemdf = pd.DataFrame(df_lista1)

steam_gamesdf.shape

user_reviewsdf.shape

user_itemdf.shape

limpieza del dataframe steam_games de datos nulos


steam_games_df1= steam_gamesdf.query('genres.notna()') #especifica que solo se deben devolver 
                                                  #las filas donde la columna genres no contiene valores nulos
steam_games_df1

Desanidamos la lista user_review

l=[] #Se crea una lista vacía para almacenar los datos del dataframe
for i in user_reviewsdf.values:#Se itera sobre cada fila del DataFrame
    l.append(i[2])#Se agrega el elemento en la posición 2 de la fila actual a la lista l

reviews_data= pd.DataFrame([review for sublist in l for review in sublist]) #Crea un nuevo DataFrame a partir de la lista de "reviews"

reviews_data.head(5)

#Se desanidan los datos de la columna "reviews"
user_reviews_sep= user_reviewsdf.explode("reviews")
user_reviews_sep

#Reseteamos los index
reviews_data.reset_index(inplace=True)
user_reviews_sep.reset_index(inplace=True)


##Unimos los Dataframe
user_riviews_1 = pd.concat([user_reviews_sep,reviews_data], axis=1) 
user_reviews_2= user_riviews_1.drop(columns = ['funny','last_edited','reviews','index','user_url' ])# Elimina columnas innecesarias
user_reviews_2.head()#previsualizamos el dataframe

Normalizamos los datos de User_items

items= [{
    'user_id': row['user_id'],
    'item_id': item['item_id'],
    'item_name': item['item_name'],
    'playtime_forever': item['playtime_forever'],
    'playtime_2weeks': item['playtime_2weeks'],
    'steam_id': row['steam_id'], 
    'items_count': row['items_count']  
} for index, row in user_itemdf.iterrows() for item in row['items']]

# Crear el DataFrame con los registros de items
items_df1 = pd.DataFrame(items)


items_df1.head()

Los convertimos en CSV

steam_games_df1.to_csv('Datasets/steam_games.csv', index=False)

user_reviews_2.to_csv('Datasets/reviews.csv', index=False)

items_df1.to_csv('Datasets/user_items.csv', index=False)
