==> 1.ETL_steam_games.ipynb <==
Importar Librerias

import json
import numpy as np
import pandas as pd

Extraer y explorar datos de nuestro csv

steam_games_df1= pd.read_csv('Datasets/steam_games.csv')

#verificamos
steam_games_df1.tail(3)

VALORES NULOS 

Buscamos valores nulos

val_null=steam_games_df1.isnull().sum()
print(val_null)

steam_games_df1.shape

Se elimina registros nulos

steam_games_df1 = steam_games_df1.dropna(how='all', subset=['id']) 

#verificamos que se elimino correctamente
val_null=steam_games_df1.isnull().sum()
print(val_null)

#vemos los tipos de datos que esta cada columna
steam_games_df1.dtypes

steam_games_df1.shape

#cambiamos el id en tipo entero ya que esta en flotate
steam_games_df1['id'] = steam_games_df1['id'].astype('Int64')

Algunas columnas estan anidadas

#Desanidamos la columna genres
steam_games_df1['genres'] = steam_games_df1['genres'].apply(eval)
lista_genres = list(set(genre for sublist in steam_games_df1['genres'] for genre in sublist))

lis=lista_genres
df_genres= pd.DataFrame(lis)
df_genres

#Aplicamos el metodo .explode para trabajar con cada valor de la lista de forma individual
steam_games_df2 = steam_games_df1.explode("genres")


#Revisamos que se realizo correctamente
steam_games_df2

steam_games_df2['release_date'] =pd.to_datetime(steam_games_df2['release_date'], errors= 'coerce')
steam_games_df2['release_date'] =steam_games_df2['release_date'].dt.year.fillna(0).astype(int)

steam_games_df2.dtypes

#Reemplazamos los valores de price
steam_games_df2.replace({"free to use": 0, "free to play": 0}, inplace=True)

steam_games_df3 = steam_games_df2.dropna(subset=['id'])
steam_games_df3.shape

steam_games_df3['id']=steam_games_df3['id'].apply(np.int64)
steam_games_df3= steam_games_df3.rename(columns={"id":"item_id"})

steam_games_df3.dtypes

steam_games_df4 = steam_games_df2.explode("genres")
steam_games_df4.shape

#Sacamos las columnas que se utilizaran para el API
steam_games_API= steam_games_df3[['item_id','app_name','title','genres','release_date','publisher','developer','price']].copy()
steam_games_API

#Sacamos las columnas que se utilizaran para el Modelo ML
steam_games_ML=steam_games_df3[['item_id','app_name','title','genres','release_date','publisher','developer']].copy()
steam_games_ML

steam_games_API.to_csv('Datasets\steam_games_API.csv', index=False)

steam_games_ML.to_csv('Datasets\steam_games_ML.csv', index=False)
