==> 2.ETL_user_reviews.ipynb <==
Importar librerias

import json
import ast
import pandas as pd


Abrir Datasets

user_reviews = pd.read_csv(r'Datasets/reviews.csv')

user_reviews.info()

user_reviews.dtypes

datos_nulos= user_reviews.isnull().sum()
datos_nulos

#Eliminar los datos nulos
user_reviews= user_reviews.dropna(subset=['posted','review','helpful','item_id'])


#Verificamos que se haya aplicado
datos_nulos= user_reviews.isnull().sum()
datos_nulos

#cambiamos el tipo de dato de la columna "recommend" "item_id" 
user_reviews['recommend'] = user_reviews['recommend'].astype('bool')
user_reviews['item_id'] = user_reviews['item_id'].astype('int64')
user_reviews.dtypes

user_reviews.head(5)

Creamos la columna sentiment_analysis

Aplicando análisis de sentimiento con NLP con la siguiente escala: debe tomar el valor '0' si es malo, '1' si es neutral y '2' si es positivo. Esta nueva columna debe reemplazar la de user_reviews.review para facilitar el trabajo de los modelos de machine learning y el análisis de datos. De no ser posible este análisis por estar ausente la reseña escrita, debe tomar el valor de 1.

Importamos liberias a utilizar

from textblob import TextBlob
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer

#Cambiamos el tipo de dato para el analisis de sentimiento pueda interpretarlo
user_reviews['review'] = user_reviews['review'].astype(str)

def analisis_sentimientos(var):
    if var is None:
        return 1 #Neutral
    analisis = TextBlob(var)
    polaridad = analisis.sentiment.polarity
    if polaridad >= 0.25:
        return 2 #Positivo
    elif polaridad <= -0.30:
        return 0 #Malo
    else:
        return 1 #Neutro

user_reviews["sentiment_analisis"]= user_reviews["review"].astype(str).apply(analisis_sentimientos)
user_reviews.head(3)

#modificamos la columna posted para que contenga solo el año posteado
user_reviews['posted'] = user_reviews['posted'].str.extract(r'Posted([\w\s\d,]+)') #se cambia la palabra Posted por espacio vacio
user_reviews['Posted'] = user_reviews['posted'].str.extract(r'(\d{4})').fillna('0') #extraigo el año y relleno los NAN por 0
user_reviews.drop(columns=['posted'])

user_reviews.to_csv(r'datasets/reviews_ETL.csv', index=False)
