==> 3.ETL_user_items.ipynb <==
Importamos librerias

import json
import numpy as np
import pandas as pd

Extraer y explorar datos de nuestro csv

items_df1= pd.read_csv('Datasets/user_items.csv')

items_df1.info()

items_df1.tail()

Buscamos valores nulos

val_null1=items_df1.isnull().sum()
print(val_null1)

#Eliminamos las columnas que no necesitamos
items_df2=items_df1.drop(columns=["playtime_2weeks"])
items_df2

items_df2.to_csv(r'Datasets\user_items_EtL.csv', index=False, encoding='utf-8')
